fix: cache check based conditional render of handleExpandItem function triggered on accordian click
This commit introduces optimizations to the event handling mechanism for expanding items in the runs
component. Previously, the event handler handleExpandItem was triggered on every accordian click.
With this change, a cache check has been implemented within the event handler to determine whether
the run or an accordian within a run being expanded already exists in the cache(done by checking the
key of the run in the cache),if the run is not found in the cache, the handleExpandItem function is
executed, thereby avoiding unnecessary network calls for runs that have already been loaded. This
optimization also handles clicking an accordian inside the runs accordian.

fix #242